requirements.txt
pygithub>=1.59.0
python-dateutil>=2.8.2
requests>=2.31.0
tabulate>=0.9.0
gitpython>=3.1.37